Adding Animation for Bus, Change Segmented Controls into Combo Menus
