+++ BSDGo/Shared/Enums/StopStatus.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum StopStatus {
+    case passed
+    case current
+    case upcoming
+}
